+++ internal/delivery/http/route/route.go
@@ -79,6 +79,7 @@ func (rc *RouteConfig) SetupExperimentRoutes(rg *gin.RouterGroup) {
 		admin.Use(middleware.AdminOnly())
 		{
 			admin.POST("", rc.ExperimentController.Create)
+			admin.POST("/:id/prepare", rc.ExperimentController.Prepare)
 		}
 	}
 }