+++ internal/infrastructure/datasource/persistance/prepared_experiment_datasource.go
@@ -1,0 +1,14 @@
+package persistance
+
+import (
+	"context"
+
+	"github.com/kshmirko/lidar-platform-go/internal/domain/entity"
+)
+
+type PreparedExperimentDataSource interface {
+	Create(ctx context.Context, exp *entity.PreparedExperiment) error
+	Update(ctx context.Context, exp *entity.PreparedExperiment) error
+	GetByID(ctx context.Context, id uint) (*entity.PreparedExperiment, error)
+	GetByExperimentID(ctx context.Context, experimentID uint) (*entity.PreparedExperiment, error)
+}